Firedrake backend: Move variables code
