+++ tlm_adjoint/firedrake/__init__.py
@@ -13,15 +13,15 @@ from .block_system import \
     (ConstantNullspace, DirichletBCNullspace, NoneNullspace, UnityNullspace)  # noqa: E402,E501,F401
 from .caches import *               # noqa: E402,F401
 from .expr import *                 # noqa: E402,F401
-from .functions import *            # noqa: E402,F401
 from .hessian_system import *       # noqa: E402,F401
 from .interpolation import *        # noqa: E402,F401
 from .parameters import *           # noqa: E402,F401
 from .projection import *           # noqa: E402,F401
 from .solve import *                # noqa: E402,F401
+from .variables import *            # noqa: E402,F401
 
 from .backend_patches import *      # noqa: E402,F401
 
 del backend, backend_interface, assembly, assignment, block_system, caches, \
-    expr, functions, hessian_system, interpolation, parameters, projection, \
-    solve, backend_patches  # noqa: F821
+    expr, hessian_system, interpolation, parameters, projection, solve, \
+    variables, backend_patches  # noqa: F821
